fixed services and controllers

+++ final-project/API/src/controllers/tennisForumController.js
@@ -24,7 +24,8 @@ router.get("/:id/", async (req, res) => {
   }
 });
 
-router.get("/comment/:commentId", async (req, res) => {
+// Get single Comment
+router.get("/comment/:commentId", isAuth, async (req, res) => {
   const commentId = req.params.commentId;
   try {
     const comment = await tennisForumService.getCommentById(commentId);
@@ -35,7 +36,7 @@ router.get("/comment/:commentId", async (req, res) => {
 });
 
 // Get Comments for a single Post
-router.get("/:id/comments", async (req, res) => {
+router.get("/:id/comments", isAuth, async (req, res) => {
   const postId = req.params.id;
   try {
     const commentsForPost = await tennisForumService.getCommentsForPost(postId);
@@ -46,15 +47,15 @@ router.get("/:id/comments", async (req, res) => {
 });
 
 // Create a Post
-router.post("/", async (req, res) => {
+router.post("/", isAuth, async (req, res) => {
   const { title, description } = req.body;
-  // const userId = req.user._id;
+  const userId = req.user._id;
 
   try {
     const createdPost = await tennisForumService.createPost({
       title,
       description,
-      postCreator: "67574fb71b9ab941005b382e",
+      postCreator: userId,
     });
 
     res.status(200).send(createdPost);
@@ -64,16 +65,16 @@ router.post("/", async (req, res) => {
 });
 
 // Create a Comment within a Post
-router.post("/:id/comment", async (req, res) => {
+router.post("/:id/comment", isAuth, async (req, res) => {
   const { content } = req.body;
   const postId = req.params.id;
-  // const userId = req.user._id;
+  const userId = req.user._id;
 
   try {
     const createComment = await tennisForumService.createComment(
       postId,
       content,
-      "67574fb71b9ab941005b382e"
+      userId
     );
 
     res.status(200).send(createComment);
@@ -83,17 +84,17 @@ router.post("/:id/comment", async (req, res) => {
 });
 
 // Updates a Post
-router.put("/:id", async (req, res) => {
+router.put("/:id", isAuth, async (req, res) => {
   const postId = req.params.id;
   const { title, description } = req.body;
 
   try {
     const post = await tennisForumService.getPostById(postId);
 
-    // if (req?.user._id != post.postCreator) {
-    //   res.status(401).send("Not Authorized");
-    //   return;
-    // }
+    if (req?.user._id != post.postCreator) {
+      res.status(401).send("Not Authorized");
+      return;
+    }
 
     const updatedPost = await tennisForumService.updatePost(
       postId,
@@ -108,17 +109,17 @@ router.put("/:id", async (req, res) => {
 });
 
 // Updates a Comment
-router.put("/comments/:id", async (req, res) => {
+router.put("/comments/:id", isAuth, async (req, res) => {
   const commentId = req.params.id;
   const { content } = req.body;
 
   try {
     const comment = await tennisForumService.getCommentById(commentId);
 
-    // if (req?.user._id != comment.commentCreator) {
-    //   res.status(401).send("Not Authorized");
-    //   return;
-    // }
+    if (req?.user._id != comment.commentCreator) {
+      res.status(401).send("Not Authorized");
+      return;
+    }
 
     const updatedPost = await tennisForumService.updateComment(
       commentId,
@@ -132,32 +133,32 @@ router.put("/comments/:id", async (req, res) => {
 });
 
 // Delete a Post
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", isAuth, async (req, res) => {
   const postId = req.params.id;
   try {
     const post = await tennisForumService.getPostById(postId);
-    // if (req?.user._id != post.postCreator) {
-    //   res.status(401).send("Not Authorized");
-    //   return;
-    // }
+    if (req?.user._id != post.postCreator) {
+      res.status(401).send("Not Authorized");
+      return;
+    }
     const deletePost = await tennisForumService.deletePost(postId);
     res.status(200).send(deletePost);
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
 });
-//TODO: add missing AUTH + encryption
+
 // Delete a Comment
-router.delete("/:postId/comments/:id", async (req, res) => {
+router.delete("/:postId/comments/:id", isAuth, async (req, res) => {
   const commentId = req.params.id;
   const postId = req.params.postId;
   try {
     const comment = await tennisForumService.getCommentById(commentId);
 
-    // if (req?.user._id != comment.commentCreator) {
-    //   res.status(401).send("Not Authorized");
-    //   return;
-    // }
+    if (req?.user._id != comment.commentCreator) {
+      res.status(401).send("Not Authorized");
+      return;
+    }
 
     const deletedComment = await tennisForumService.deleteComment(
       commentId,
